@@ -1,15 +1,17 @@
 import { Injectable } from '@angular/core';
-import { HttpClient,HttpParams } from '@angular/common/http';
+import { HttpClient, HttpParams } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { Item } from './item.model';
-import { Category } from './category.model';
+import { Item } from '../models/item.model';
+// import { Category } from './category.model';
+import { environment } from '../../environments/environment';
+
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class ItemService {
-  private baseUrl = 'http://localhost:8082/api/v1/item';
+  private readonly baseUrl = environment.itemApiUrl;
 
-  constructor(private http: HttpClient) {}
+  constructor(private readonly http: HttpClient) {}
 
   getItems(): Observable<Item[]> {
     return this.http.get<Item[]>(`${this.baseUrl}/all`);
@@ -18,38 +20,38 @@ export class ItemService {
   createItem(item: Item): Observable<Item> {
     return this.http.post<Item>(`${this.baseUrl}/new`, item);
   }
-  
+
   getItemById(id: number): Observable<Item> {
-    return this.http.get<Item>(`${this.baseUrl}/${id}`); 
+    return this.http.get<Item>(`${this.baseUrl}/${id}`);
   }
-  
-  filterItems(categoryIds?: number[], minPrice?: number, maxPrice?: number): Observable<Item[]> {
+
+  filterItems(
+    categoryIds?: number[],
+    minPrice?: number,
+    maxPrice?: number
+  ): Observable<Item[]> {
     let params = new HttpParams();
 
     // Agrega las categorías solo si `categoryIds` no está vacío
-    if (categoryIds !== undefined){
-    if (categoryIds.length > 0) {
-        categoryIds.forEach(id => {
-            params = params.append('categoryIds', id);
+    if (categoryIds !== undefined) {
+      if (categoryIds.length > 0) {
+        categoryIds.forEach((id) => {
+          params = params.append('categoryIds', id);
         });
-    }}
+      }
+    }
 
     // Añade el precio mínimo si está definido
     if (minPrice !== undefined) {
-        params = params.set('minPrice', minPrice.toString());
+      params = params.set('minPrice', minPrice.toString());
     }
 
     // Añade el precio máximo si está definido
     if (maxPrice !== undefined) {
-        params = params.set('maxPrice', maxPrice.toString());
+      params = params.set('maxPrice', maxPrice.toString());
     }
     console.log(params);
     // Realiza la solicitud GET con los parámetros
     return this.http.get<Item[]>(`${this.baseUrl}/filter`, { params });
-}
-
-
-
-
-
+  }
 }
